feat: make blockchain JSON RPC configurable from settings
