+++ src/utils/net.ts
@@ -6,8 +6,8 @@ export function getJson<T extends Record<string, any>>(url: string): Promise<T> 
   return sendRequest(url, 'GET') as Promise<T>
 }
 
-export function postJson(url: string, data?: Record<string, any>): Promise<Record<string, unknown>> {
-  return sendRequest(url, 'POST', data)
+export function postJson<T extends Record<string, any>>(url: string, data?: T): Promise<T> {
+  return sendRequest(url, 'POST', data) as Promise<T>
 }
 
 export async function sendRequest(
